Change indices dtype to integers

+++ bw_processing/constants.py
@@ -18,7 +18,7 @@ UNCERTAINTY_DTYPE = [
     ("maximum", np.float32),
     ("negative", np.bool),
 ]
-INDICES_DTYPE = [("row", np.float32), ("col", np.float32)]
+INDICES_DTYPE = [("row", np.int32), ("col", np.int32)]
 
 NAME_RE = re.compile(r"^[\w\-\.]*$")
